Handle regexps in .bzrignore-upload and take parents into account.
* tests/test_upload.py:
(UploadUtilsMixin._get_cmd_upload): The hell with bzrlib
regressions, just use sys.stdout.
(TestUploadMixin.test_ignore_regexp): Ensure regexps are handled.
(TestUploadMixin.test_ignore_change_file_into_dir,
TestUploadMixin.test_ignore_change_dir_into_file,
TestUploadMixin.test_ignore_delete_dir_in_subdir,
TestIncrementalUpload.test_ignore_delete_one_file): More ignore tests.

* __init__.py:
(BzrUploader._get_ignored, BzrUploader.is_ignored): Handle regexps
and ensure no parents are ignored.
(BzrUploader.upload_file, BzrUploader.make_remote_dir,
BzrUploader.delete_remote_dir, BzrUploader.delete_remote_file):
Use is_ignored.
(BzrUploader.upload_full_tree, BzrUploader.upload_tree): Use
is_ignored.

+++ __init__.py
@@ -164,6 +164,7 @@ import stat
 from bzrlib import (
     bzrdir,
     errors,
+    globbing,
     ignores,
     revisionspec,
     transport,
@@ -262,6 +263,7 @@ class BzrUploader(object):
         self._pending_deletions = []
         self._pending_renames = []
         self._uploaded_revid = None
+        self._ignored = None
 
     def set_uploaded_revid(self, rev_id):
         # XXX: Add tests for concurrent updates, etc.
@@ -279,18 +281,31 @@ class BzrUploader(object):
                 self._uploaded_revid = revision.NULL_REVISION
         return self._uploaded_revid
 
-    def get_ignored(self):
-        """Get upload-specific ignored files from the current branch"""
-        try:
-            ignore_file = self.tree.get_file_by_path('.bzrignore-upload')
-            return ignores.parse_ignore_file(ignore_file)
-        except errors.NoSuchId:
-            return []
+    def _get_ignored(self):
+        if self._ignored is None:
+            try:
+                ignore_file = self.tree.get_file_by_path('.bzrignore-upload')
+                ignored_patterns = ignores.parse_ignore_file(ignore_file)
+            except errors.NoSuchId:
+                ignored_patterns = []
+            self._ignored = globbing.Globster(ignored_patterns)
+        return self._ignored
+
+    def is_ignored(self, relpath):
+        glob = self._get_ignored()
+        ignored = glob.match(relpath)
+        import os
+        if not ignored:
+            # We still need to check that all parents are not ignored
+            dir = os.path.dirname(relpath)
+            while dir and not ignored:
+                ignored = glob.match(dir)
+                if not ignored:
+                    dir = os.path.dirname(dir)
+        return ignored
 
     def upload_file(self, relpath, id, mode=None):
-        ignored_files = self.get_ignored()
-        is_ignored =  osutils.is_inside_any(ignored_files, relpath)
-        if is_ignored is False:
+        if not self.is_ignored(relpath):
             if mode is None:
                 if self.tree.is_executable(id):
                     mode = 0775
@@ -298,7 +313,8 @@ class BzrUploader(object):
                     mode = 0664
             if not self.quiet:
                 self.outf.write('Uploading %s\n' % relpath)
-            self.to_transport.put_bytes(relpath, self.tree.get_file_text(id), mode)
+            self.to_transport.put_bytes(relpath, self.tree.get_file_text(id),
+                                        mode)
         else:
             if not self.quiet:
                 self.outf.write('Ignoring %s\n' % relpath)
@@ -322,9 +338,7 @@ class BzrUploader(object):
         self.upload_file(relpath, id, mode)
 
     def make_remote_dir(self, relpath, mode=None):
-        ignored_files = self.get_ignored()
-        is_ignored =  osutils.is_inside_any(ignored_files, relpath)
-        if is_ignored is False:
+        if not self.is_ignored(relpath):
             if mode is None:
                 mode = 0775
             self.to_transport.mkdir(relpath, mode)
@@ -353,14 +367,24 @@ class BzrUploader(object):
         self.make_remote_dir(relpath, mode)
 
     def delete_remote_file(self, relpath):
-        if not self.quiet:
-            self.outf.write('Deleting %s\n' % relpath)
-        self.to_transport.delete(relpath)
+        if not self.is_ignored(relpath):
+            if not self.quiet:
+                self.outf.write('Deleting %s\n' % relpath)
+            self.to_transport.delete(relpath)
+        else:
+            if not self.quiet:
+                self.outf.write('Ignoring %s\n' % relpath)
 
     def delete_remote_dir(self, relpath):
-        if not self.quiet:
-            self.outf.write('Deleting %s\n' % relpath)
-        self.to_transport.rmdir(relpath)
+        if not self.is_ignored(relpath):
+            if not self.quiet:
+                self.outf.write('Deleting %s\n' % relpath)
+            self.to_transport.rmdir(relpath)
+        else:
+            # XXX: Add a test where a subdir is ignored but we still want to
+            # delete the dir -- vila 100106
+            if not self.quiet:
+                self.outf.write('Ignoring %s\n' % relpath)
 
     def delete_remote_dir_maybe(self, relpath):
         """Try to delete relpath, keeping failures to retry later."""
@@ -418,8 +442,12 @@ class BzrUploader(object):
             for relpath, ie in self.tree.inventory.iter_entries():
                 if relpath in ('', '.bzrignore', '.bzrignore-upload'):
                     # skip root ('')
-                    # .bzrignore has no meaning outside of a working tree
-                    # so do not upload it
+                    # .bzrignore and .bzrignore-upload have no meaning outside
+                    # a working tree so do not upload them
+                    continue
+                if self.is_ignored(relpath):
+                    if not self.quiet:
+                        self.outf.write('Ignoring %s\n' % relpath)
                     continue
                 if ie.kind == 'file':
                     self.upload_file_robustly(relpath, ie.file_id)
@@ -458,6 +486,10 @@ class BzrUploader(object):
         self.tree.lock_read()
         try:
             for (path, id, kind) in changes.removed:
+                if self.is_ignored(path):
+                    if not self.quiet:
+                        self.outf.write('Ignoring %s\n' % path)
+                    continue
                 if kind is 'file':
                     self.delete_remote_file(path)
                 elif kind is  'directory':
@@ -467,6 +499,11 @@ class BzrUploader(object):
 
             for (old_path, new_path, id, kind,
                  content_change, exec_change) in changes.renamed:
+                if self.is_ignored(old_path) and self.is_ignored(new_path):
+                    if not self.quiet:
+                        self.outf.write('Ignoring %s\n' % old_path)
+                        self.outf.write('Ignoring %s\n' % new_path)
+                    continue
                 if content_change:
                     # We update the old_path content because renames and
                     # deletions are differed.
@@ -476,6 +513,10 @@ class BzrUploader(object):
             self.finish_deletions()
 
             for (path, id, old_kind, new_kind) in changes.kind_changed:
+                if self.is_ignored(path):
+                    if not self.quiet:
+                        self.outf.write('Ignoring %s\n' % path)
+                    continue
                 if old_kind is 'file':
                     self.delete_remote_file(path)
                 elif old_kind is  'directory':
@@ -491,6 +532,10 @@ class BzrUploader(object):
                     raise NotImplementedError
 
             for (path, id, kind) in changes.added:
+                if self.is_ignored(path):
+                    if not self.quiet:
+                        self.outf.write('Ignoring %s\n' % path)
+                    continue
                 if kind is 'file':
                     self.upload_file(path, id)
                 elif kind is 'directory':
@@ -501,6 +546,10 @@ class BzrUploader(object):
             # XXX: Add a test for exec_change
             for (path, id, kind,
                  content_change, exec_change) in changes.modified:
+                if self.is_ignored(path):
+                    if not self.quiet:
+                        self.outf.write('Ignoring %s\n' % path)
+                    continue
                 if kind is 'file':
                     self.upload_file(path, id)
                 else:
